Move the "to_domain" func to "o1_utils::evaluations" file
